requirements.txt
Django==1.8.2
Markdown==2.6.2
djangorestframework==3.1.3
psycopg2==2.6.1
Pillow